backend/requirements.txt
amqp==5.0.2
beautifulsoup4==4.9.3
billiard==3.6.3.0
celery==5.0.5
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
Django==2.2
django-ckeditor==6.0.0
django-cors-headers==3.6.0
django-ddp==0.19.1
django-debug-toolbar==3.2
django-js-asset==1.2.2
django-phonenumber-field==5.0.0
djangorestframework==3.12.2
gevent==20.12.0
gevent-websocket==0.10.1
greenlet==0.4.17
idna==2.10
importlib-metadata==3.3.0
kombu==5.0.2
meteor-ejson==1.1.0
phonenumbers==8.12.15
Pillow==8.0.1
prompt-toolkit==3.0.8
psycogreen==1.0.2
psycopg2==2.8.6
pybars3==0.9.7
pycparser==2.20
PyJWT==2.0.0
PyMeta3==0.5.1
pytz==2020.4
requests==2.25.1
six==1.15.0
soupsieve==2.1
sqlparse==0.4.1
twilio==6.50.1
typing-extensions==3.7.4.3
urllib3==1.26.2
vine==5.0.0
wcwidth==0.2.5
zipp==3.4.0
zope.event==4.5.0
zope.interface==5.2.0
